refactor(guard): move the test-only registration under testing/
Addresses three review threads.

The test-only registration path moves out of registry.ts, which every
production import graph pulls in, into src/testing/register.ts. The slot
primitives it shares with the registry move to src/registration-slot.ts,
so neither module has to export something only tests use. Verified by
walking the built graph: dist/node.js, dist/fetch.js and dist/bun.js
reach no module under dist/testing/, and registerArcjetForTesting no
longer appears in dist/registry.js.

The test client itself moves to src/testing/index.ts, so the ./testing
subpath now resolves to dist/testing/index.js. Its tests split along the
same seams, with the shared recorder doubles under test/_shared/ because
the build emits everything in src/ that is not a *.test.ts.

The README testing example now leads with `using` and keeps the
Node.js 22 form in a details block, rather than the other way around.

Both example forms were also racy: refund() is async, so an unawaited
call let the assertion run before the code under test reached capture().
The callbacks are async and the call is awaited. The surrounding prose
said assertions "need no waiting", which read as though that applied to
the code under test; it now says recording is synchronous once capture()
is reached.
