+++ arcjet-guard/jsr.json
@@ -6,7 +6,7 @@
     ".": "./src/fetch.ts",
     "./node": "./src/node.ts",
     "./fetch": "./src/fetch.ts",
-    "./testing": "./src/testing.ts"
+    "./testing": "./src/testing/index.ts"
   },
   "publish": {
     "include": ["src/**/*.ts", "src/**/*.js", "package.json", "LICENSE", "README.md"]
